Refactor and rearrange some code
* Refactor examples
* Move Laplacians to the same file
* Add a guard for C++ compiler
* Minor changes
* Cleanup some includes
* Move some iterative methods from src/precond/ to src/
