+++ src/genmap-laplacian.c
@@ -43,7 +43,7 @@ static void genmap_find_neighbors(struct array *nbrs, genmap_handle h,
   struct array a;
   array_init(csr_entry, &a, 10);
 
-  // FIXME: Assumes quads or hexes
+  /* FIXME: Assumes quads or hexes */
   sint s = 0, e;
   csr_entry t;
   while (s < size) {
@@ -64,9 +64,7 @@ static void genmap_find_neighbors(struct array *nbrs, genmap_handle h,
   }
 
   sarray_transfer(csr_entry, &a, proc, 1, &cr);
-  // TODO: Check if the last line is redundant
   sarray_sort_2(csr_entry, a.ptr, a.n, r, 1, c, 1, &h->buf);
-  // sarray_sort(csr_entry, a.ptr, a.n, r, 1, &h->buf);
 
   array_init(entry, nbrs, lelt);
 
@@ -116,6 +114,89 @@ int GenmapInitLaplacian(genmap_handle h, struct comm *c) {
 int GenmapLaplacian(genmap_handle h, GenmapScalar *u, GenmapScalar *v) {
   csr_mat_gather(h->M, h->gsh, u, h->b, &h->buf);
   csr_mat_apply(v, h->M, h->b);
+
+  return 0;
+}
+
+int GenmapInitLaplacianWeighted(genmap_handle h, struct comm *c) {
+  GenmapInt lelt = genmap_get_nel(h);
+  GenmapInt nv = genmap_get_nvertices(h);
+
+  GenmapRealloc(lelt, &h->weights);
+  GenmapUInt numPoints = (GenmapUInt)nv * lelt;
+
+  GenmapLong *vertices;
+  GenmapMalloc(numPoints, &vertices);
+
+  struct rsb_element *elements = genmap_get_elements(h);
+  GenmapInt i, j;
+  for (i = 0; i < lelt; i++) {
+    for (j = 0; j < nv; j++)
+      vertices[i * nv + j] = elements[i].vertices[j];
+  }
+
+  if (h->gsw != NULL)
+    gs_free(h->gsw);
+
+#if defined(GENMAP_DEBUG)
+  double t1 = GenmapGetMaxRss();
+  if (c->id == 0)
+    printf("RSS before gs_setup: %lf\n", t1);
+#endif
+
+  h->gsw = gs_setup(vertices, numPoints, c, 0, gs_crystal_router, 0);
+
+#if defined(GENMAP_DEBUG)
+  t1 = GenmapGetMaxRss();
+  if (c->id == 0)
+    printf("RSS after gs_setup: %lf\n", t1);
+#endif
+
+  GenmapScalar *u;
+  GenmapMalloc(numPoints, &u);
+
+  for (i = 0; i < lelt; i++)
+    for (j = 0; j < nv; j++)
+      u[nv * i + j] = 1.;
+
+  gs(u, gs_double, gs_add, 0, h->gsw, &h->buf);
+
+  for (i = 0; i < lelt; i++) {
+    h->weights[i] = 0.0;
+    for (j = 0; j < nv; j++)
+      h->weights[i] += u[nv * i + j];
+  }
+
+  for (i = 0; i < lelt; i++)
+    h->weights[i] *= -1;
+
+  GenmapFree(u);
+  GenmapFree(vertices);
+
+  return 0;
+}
+
+int GenmapLaplacianWeighted(genmap_handle h, GenmapScalar *u, GenmapScalar *v) {
+  GenmapInt lelt = genmap_get_nel(h);
+  GenmapInt nv = genmap_get_nvertices(h);
+
+  GenmapScalar *ucv;
+  GenmapMalloc((size_t)(nv * lelt), &ucv);
+
+  GenmapInt i, j;
+  for (i = 0; i < lelt; i++)
+    for (j = 0; j < nv; j++)
+      ucv[nv * i + j] = u[i];
+
+  gs(ucv, gs_double, gs_add, 0, h->gsw, &h->buf);
+
+  for (i = 0; i < lelt; i++) {
+    v[i] = h->weights[i] * u[i];
+    for (j = 0; j < nv; j++)
+      v[i] += ucv[nv * i + j];
+  }
+
+  GenmapFree(ucv);
 
   return 0;
 }
